protect contacts routes

+++ controllers/contactController.js
@@ -2,15 +2,15 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async(req,res)=>{
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts);
 })
 
 //@desc Get contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async(req,res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
@@ -22,7 +22,7 @@ const getContact = asyncHandler(async(req,res)=>{
 
 //@desc Create contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req,res)=>{
     console.log("The request body is: ", req.body);
     const {name, email, phone} = req.body;
@@ -31,14 +31,14 @@ const createContact = asyncHandler(async (req,res)=>{
         throw new Error("All fields are mandatory !");
     }
     const contact = await Contact.create({
-        name,email,phone,
+        name,email,phone,user_id: req.user.id,
     });
     res.status(201).json(contact);
 })
 
 //@desc Update contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async(req,res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
@@ -46,6 +46,10 @@ const updateContact = asyncHandler(async(req,res)=>{
         throw new Error("Contact not found");
     }
 
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User dont have permission to other users contact");
+    }
     const updateContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -56,14 +60,17 @@ const updateContact = asyncHandler(async(req,res)=>{
 
 //@desc Delete contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async(req,res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         req.status(404);
         throw new Error("Contact not found");
     }
-
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User dont have permission to other users contact");
+    }
     await contact.deleteOne();
     res.status(200).json(contact);
 })
